little update, mainly fix compiler error

+++ InnerDetector/InDetConditions/PixelConditionsData/PixelConditionsData/PixelByteStreamErrors.h
@@ -22,7 +22,7 @@ namespace PixelByteStreamErrors {
   //!< @brief helper to set the error: @example errors[hashId] = addError( errors[hashId], PixelByteStreamErrors::Invalid )
   inline void addError(IDCInDetBSErrContainer::ErrorCode& errWord,  PixelErrorsEnum errType ) { errWord |= makeError( errType ); }
 
-  inline void addServiceRecords(IDCInDetBSErrContainer::ErrorCode& errWord, int errorCount ) { errWord = errorCount; }
+  inline void addServiceRecords(IDCInDetBSErrContainer::ErrorCode& errWord, uint64_t errorCount ) { errWord = errorCount; }
 
 }
